requirements_new_utf8.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
av==16.1.0
botocore==1.42.67
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
cryptography==46.0.5
distro==1.9.0
exceptiongroup==1.3.1
fasttext-predict==0.9.2.4
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
google-auth==2.49.1
google-genai==1.70.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-retries==0.4.6
humanfriendly==10.0
idna==3.11
ImageIO==2.37.3
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
lazy-loader==0.5
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modelscope==1.34.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.0
onnxruntime==1.23.2
openai==2.26.0
opencv-python==4.13.0.92
packaging==26.0
pdfminer.six==20260107
pdftext==0.6.3
pillow==12.1.1
protobuf==7.34.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.27.2
pypdf==6.8.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
qwen-vl-utils==0.0.14
reportlab==4.4.10
requests==2.32.5
rich==14.3.3
robust-downloader==0.0.2
s3transfer==0.16.0
scikit-image==0.25.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
tenacity==9.1.4
tifffile==2025.5.10
torch==2.10.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
win32_setctime==1.2.0

